Add main layout and routing system
