+++ package.json
@@ -95,6 +95,7 @@
     "postcss": "^8.4.32",
     "postcss-loader": "^7.3.3",
     "prettier": "^3.0.3",
+    "react-router-dom": "^6.21.1",
     "tailwindcss": "^3.4.0",
     "ts-jest": "^29.1.1",
     "ts-loader": "^9.4.4",
