--- src/components/Popup/PopupFooter/PopupFooter.scss
@@ -1,11 +1,0 @@
-.footer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0 20px;
-  border-top: 1px solid #e2e8f0;
-  color: #4a5568;
-  flex-basis: 48px;
-  flex-shrink: 0;
-  font-size: 14px;
-}
