--- src/components/Popup/PopupFooter/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PopupFooter'
