--- src/components/Popup/PopupHeader/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PopupHeader'
