+++ src/popup/pages/Settings.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export const Settings = () => {
+  return <>Settings</>
+}
